refactor(fluid-cursor): merge FluidCursorAdvanced into FluidCursor
FluidCursorAdvanced was a superset of FluidCursor. Merge all features
(autoSplat, interactive, pauseWhenHidden, splatOnMount, allowMultiple,
singleton management) into FluidCursor with a new `contained` prop
(default true). FluidCursorAdvanced is re-exported as a deprecated
alias for backwards compatibility.

BREAKING CHANGE: FluidCursor now defaults to contained mode (absolute
positioning). Pass `contained={false}` for previous fullscreen behavior.
