+++ .changeset/merge-fluid-cursor.md
@@ -1,0 +1,5 @@
+---
+"fancy-ui-svelte": major
+---
+
+refactor(fluid-cursor): merge FluidCursorAdvanced into FluidCursor with new `contained` prop
