+++ src/lib/fancy-ui/fluid-cursor/FluidCursor.test.ts
@@ -17,27 +17,30 @@ describe("FluidCursor", () => {
 		expect(canvas).toBeInTheDocument();
 	});
 
-	it('canvas has id="fluid"', () => {
-		const { container } = render(FluidCursor);
-		const canvas = container.querySelector("canvas#fluid");
-		expect(canvas).toBeInTheDocument();
-	});
-
 	it("container has pointer-events-none class", () => {
 		const { container } = render(FluidCursor);
 		const div = container.firstElementChild as HTMLElement;
 		expect(div?.className).toContain("pointer-events-none");
 	});
 
-	it("container has fixed positioning class", () => {
+	it("defaults to contained mode (absolute positioning)", () => {
 		const { container } = render(FluidCursor);
 		const div = container.firstElementChild as HTMLElement;
-		expect(div?.className).toContain("fixed");
+		expect(div?.className).toContain("absolute");
+		expect(div?.className).toContain("inset-0");
 	});
 
-	it("container has z-50 class", () => {
+	it("canvas fills container in contained mode", () => {
 		const { container } = render(FluidCursor);
+		const canvas = container.querySelector("canvas") as HTMLElement;
+		expect(canvas?.className).toContain("h-full");
+		expect(canvas?.className).toContain("w-full");
+	});
+
+	it("uses fixed positioning when contained={false}", () => {
+		const { container } = render(FluidCursor, { props: { contained: false } });
 		const div = container.firstElementChild as HTMLElement;
+		expect(div?.className).toContain("fixed");
 		expect(div?.className).toContain("z-50");
 	});
 
@@ -45,6 +48,34 @@ describe("FluidCursor", () => {
 		const { container } = render(FluidCursor, { props: { class: "my-fluid" } });
 		const div = container.firstElementChild as HTMLElement;
 		expect(div?.className).toContain("my-fluid");
+	});
+
+	describe("interaction props", () => {
+		it("interactive={false} does not register mouse event listeners", () => {
+			const addSpy = vi.spyOn(window, "addEventListener");
+			render(FluidCursor, { props: { interactive: false } });
+			const mousedownCalls = addSpy.mock.calls.filter(([event]) => event === "mousedown");
+			expect(mousedownCalls).toHaveLength(0);
+			const mousemoveCalls = addSpy.mock.calls.filter(([event]) => event === "mousemove");
+			expect(mousemoveCalls).toHaveLength(0);
+			addSpy.mockRestore();
+		});
+
+		it("mounts without error with autoSplat={true} and splatOnMount={true}", () => {
+			const { container } = render(FluidCursor, {
+				props: { autoSplat: true, splatOnMount: true },
+			});
+			expect(container.querySelector("canvas")).toBeInTheDocument();
+		});
+
+		it("pauseWhenHidden={false} does not instantiate IntersectionObserver", () => {
+			const observeSpy = vi.fn();
+			const mockObserver = vi.fn(() => ({ observe: observeSpy, disconnect: vi.fn() }));
+			vi.stubGlobal("IntersectionObserver", mockObserver);
+			render(FluidCursor, { props: { pauseWhenHidden: false } });
+			expect(mockObserver).not.toHaveBeenCalled();
+			vi.unstubAllGlobals();
+		});
 	});
 
 	describe("color props", () => {
@@ -61,7 +92,6 @@ describe("FluidCursor", () => {
 		});
 
 		it("renders without error when fluidColor and fluidColors are both provided (fluidColor takes priority)", () => {
-			// Both props accepted without throw — fluidColor priority is exercised at runtime
 			const { container } = render(FluidCursor, {
 				props: { fluidColor: "#ff0000", fluidColors: ["#00ffcc", "#7700ff"] },
 			});
